Fetch the active (or all) gene sets from a GeneSetDb or SparrowResult

Fetch the active (or all) gene sets from a GeneSetDb or SparrowResult

geneSets(x, ...)

# S4 method for GeneSetDb
length(x)

# S4 method for GeneSetDb
geneSets(x, active.only = is.conformed(x), ..., as.dt = FALSE)

# S4 method for GeneSetDb
nrow(x)

# S4 method for SparrowResult
geneSets(x, ..., as.dt = FALSE)

Arguments

x

Object to retrieve the gene set from, either a GeneSetDb or a SparrowResult.

...

pass through arguments

active.only

only look for gene sets that are "active"? Defaults to TRUE if x is conformed to a target expression object, else FALSE. conform() for further details.

as.dt

If FALSE (default), the data.frame like thing that this funciton returns will be set to a data.frame. Set this to TRUE to keep this object as a data.table

Value

a data.table with geneset information.

Methods (by class)

GeneSetDb: Returns the number of genesets in a GeneSetDb

GeneSetDb: return all genesets from a GeneSetDb

GeneSetDb: return number of genesets in GeneSetDb

SparrowResult: return the active genesets from a SparrowResult

Examples

gdb <- exampleGeneSetDb()
gs <- geneSets(gdb)