Assembles a matrix of nominal or adjusted pvalues from a sparrow::seas result

You might want a matrix of pvalues (or FDRs) for the gene sets across all GSEA methods you tried. I think I did, once, so here it is.

p.matrix(
  x,
  names = resultNames(x),
  pcol = c("padj", "padj.by.collection", "pval")
)

Arguments

x

A SparrowResult() object.

names

the entries from resultNames(x) that you want to include in the matrix. By default we take all of them.

pcol

The name of the column in logFC(x) where the type of pvalues are that we are collection. Pick on of "padj", "padj.by.collection", or "pval"

Value

A matrix of the desired pvalues for all genesets

Examples

mg <- exampleSparrowResult()
pm <- p.matrix(mg)