Simulate gene expression data

This function simulates gene expression data using a spot-based model

Usage

simulate_data(
  n,
  nct,
  effect_range,
  intercept_range,
  min_effect = 0,
  library_size = 500,
  spot_ct = min(2, nct),
  p = 6,
  num_null = 2,
  prob_ct = NULL,
  family = "poisson",
  dispersion = 1
)

Arguments

n

Number of data points

nct

Number of cell types

effect_range

Range of effect sizes (effect sizes drawn uniformly on this interval)

intercept_range

Range for intercepts (intercept sizes drawn uniformly on this interval)

min_effect

Minimum absolute value of effect sizes.

library_size

Number of transcripts per cell/spot

spot_ct

Number of cell types per spot

p

Number of covariates

num_null

Number of null coefficients to include

prob_ct

Optional probabilities for cell type sampling

family

The data generating distribution. One of poisson, negative binomial,binomial, and gaussian

dispersion

Dispersion parameter for NB or Gaussian family. Size parameter for NB and sd for Gaussian. Default: 1

Value

A list with simulated y, X, lambda, beta, null_beta, and CT