Match a species query to the regularized species info.

Match a species query to the regularized species info.

species_info(query = NULL, ...)

Arguments

query

the species name to lookup, if NULL (default), returns the internal species info table, otherwise the row of the table that matches query.

...

pass through

Value

a data.frame of species-related information that is used to fetch appropriate annotation files and conversion functions between species for gene identifiers, and such.

Examples

species_info()
#>        alias                  species bioc_full bioc_abbrev kegg
#> 1                                            NA          Ag  aga
#> 2                                            NA          At  ath
#> 3                                            NA          Bt  bta
#> 4       worm   Caenorhabditis elegans        NA          Ce  cel
#> 5        fly  Drosophila melanogaster        NA          Dm  dme
#> 6  zebrafish              Danio rerio        NA          Dr  dre
#> 7                                            NA       EcK12  eco
#> 8                                            NA     EcSakai  ecs
#> 9                                            NA          Gg  gga
#> 10     human             Homo sapiens        NA          Hs  hsa
#> 11      cyno      Macaca fascicularis        NA          Mf  mcf
#> 12     mouse             Mus musculus        NA          Mm  mmu
#> 13    rhesus           Macaca mulatta        NA         Mmu  mcc
#> 14                                           NA          Pf  pfa
#> 15                                           NA          Pt  ptr
#> 16       rat        Rattus norvegicus        NA          Rn  rno
#> 17     yeast Saccharomyces cerevisiae        NA          Ss  ssc
#> 18                                           NA          Xl  xla
#>                  mart_dataset                symbol_name
#> 1                                                       
#> 2                                                       
#> 3                                                       
#> 4                                                       
#> 5            flybasename_gene dmelanogaster_gene_ensembl
#> 6         drerio_gene_ensembl             zfin_id_symbol
#> 7                                                       
#> 8                                                       
#> 9                                                       
#> 10      hsapiens_gene_ensembl                hgnc_symbol
#> 11 mfascicularis_gene_ensembl         external_gene_name
#> 12     mmusculus_gene_ensembl                 mgi_symbol
#> 13                                                      
#> 14                                                      
#> 15                                                      
#> 16                                                      
#> 17                                                      
#> 18                                                      
species_info("human")
#> $alias
#> [1] "human"
#> 
#> $species
#> [1] "Homo sapiens"
#> 
#> $bioc_full
#> [1] NA
#> 
#> $bioc_abbrev
#> [1] "Hs"
#> 
#> $kegg
#> [1] "hsa"
#> 
#> $mart_dataset
#> [1] "hsapiens_gene_ensembl"
#> 
#> $symbol_name
#> [1] "hgnc_symbol"
#>